refactor: move screen to its own folder
